+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 * Added initial support for JavaFX 2
 * Added EqualsBooleanRule in core
+* Added convenient BooleanResultCollector in core
+* Added IsTrueRule and IsFalseRule in core
 * Added convenient constructor to IconComponentDecoration
 * Fixed NullPointerException when calling dispose() multiple times on some triggers